backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML
google-generativeai==0.3.2
sentence-transformers==2.3.1
chromadb==0.4.22

# Data & DB
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Web Scraping
beautifulsoup4==4.12.3
playwright==1.41.0
httpx==0.26.0

# Utilities
python-multipart==0.0.6
pandas==2.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3